requirements.txt
ccxt
fastapi
uvicorn
sqlalchemy
pandas
aiohttp
python-dotenv
pydantic
pytest
httpx
vaderSentiment
feedparser
matplotlib>=3.9
scikit-learn>=1.4
joblib>=1.3
pyyaml
pyarrow
